feat: navigates forward to place detail screen when list item is selected
