+++ App.js
@@ -4,6 +4,7 @@ import { Provider } from "react-redux";
 import AuthScreen from "./src/screens/Auth/Auth";
 import SharePlaceScreen from "./src/screens/SharePlace/SharePlace";
 import FindPlaceScreen from "./src/screens/FindPlace/FindPlace";
+import PlaceDetailScreen from "./src/screens/PlaceDetail/PlaceDetail";
 
 import configureStore from "./src/store/configureStore";
 
@@ -27,6 +28,10 @@ Navigation.registerComponent(
   () => FindPlaceScreen,
   store,
   Provider
+);
+Navigation.registerComponent(
+  "awesome-places.PlaceDetailScreen",
+  () => PlaceDetailScreen
 );
 
 // Start a navigation app
