+++ src/screens/FindPlace/FindPlace.js
@@ -5,10 +5,27 @@ import { connect } from "react-redux";
 import PlaceList from "../../components/PlaceList/PlaceList";
 
 class FindPlaceScreen extends Component {
+  itemSelectedHandler = key => {
+    const selectedPlace = this.props.places.find(place => {
+      return place.key === key;
+    });
+
+    this.props.navigator.push({
+      screen: "awesome-places.PlaceDetailScreen",
+      title: selectedPlace.name,
+      passProps: {
+        selectedPlace: selectedPlace
+      }
+    });
+  };
+
   render() {
     return (
       <View>
-        <PlaceList places={this.props.places} />
+        <PlaceList
+          places={this.props.places}
+          onItemSelected={this.itemSelectedHandler}
+        />
       </View>
     );
   }
